Feat: Process Offline Queued CRUD Actions

+++ providers/FoldersProvider.tsx
@@ -7,6 +7,9 @@ import { createFolder, deleteFolder } from "../services/folder";
 import { FetchingStatusType } from "../types/fetchingStatus";
 import { ErrorType } from "../types/errorType";
 import { cacheFolder, readAllFoldersCache, removeFolderCache, syncFoldersCache } from "../storage/folder";
+import { useDeviceContext } from "../context/DeviceContext";
+import { FolderMutation } from "../storage/types";
+import { dequeueMutation, enqueueMutation, readMutationQueue } from "../storage/folderQueue";
 
 export type SuccessType = {
   error: false;
@@ -16,8 +19,9 @@ export type SuccessType = {
 
 export default function FoldersProvider({children} : PropsWithChildren) {
   const { userId } = useAuthContext()
-  const [userFolders, setUserFolders] = useState<Folder[] | []>([]);
+  const [userFolders, setUserFolders] = useState<Folder[]>([]);
   const [folderFetchingStatus, setFolderFetchingStatus] = useState<FetchingStatusType>('idle');
+  const { networkStatus } = useDeviceContext();
 
   async function refreshFolders(): Promise<void> {
     // Read from cache
@@ -28,18 +32,20 @@ export default function FoldersProvider({children} : PropsWithChildren) {
     }
 
     // Read from backend
-    if (!userId) return;
-    setFolderFetchingStatus('loading')
-    try {
-      const data = await getFolders(userId);
-      if(data) {
-        setFolderFetchingStatus('success');
-        setUserFolders(data);
-        await syncFoldersCache(data);
+    if(networkStatus === 'online') {
+      if (!userId) return;
+      setFolderFetchingStatus('loading')
+      try {
+        const data = await getFolders(userId);
+        if(data) {
+          setFolderFetchingStatus('success');
+          setUserFolders(data);
+          await syncFoldersCache(data);
+        }
+      } catch (err) {
+        setFolderFetchingStatus('error')
+        console.log("Failed to fetch folders:", err);
       }
-    } catch (err) {
-      setFolderFetchingStatus('error')
-      console.log("Failed to fetch folders:", err);
     }
   }
 
@@ -49,10 +55,38 @@ export default function FoldersProvider({children} : PropsWithChildren) {
       messageTitle: 'Failed to add folder'
     };
     try {
+      if(networkStatus === 'offline') {
+        const tempId = `temp-${Date.now()}`;
+
+        const localFolder: Folder = {
+          id: tempId,
+          created_at: Date.now().toString(),
+          name: folderName,
+          user_id: userId,
+        }
+
+        const mutation: FolderMutation = {
+          id: tempId,
+          type: 'ADD_FOLDER',
+          tempId: localFolder.created_at,
+          payload: {name: folderName},
+        }
+
+        await enqueueMutation(mutation);
+        setUserFolders(prev => [...prev, localFolder]);
+        await cacheFolder(localFolder.id, localFolder);
+
+        return {
+          error: false,
+          message: 'Folder queued for addition.'
+        }
+      }
+
       const data = await createFolder(folderName, userId);
       if(!data) {
         throw new Error("Folder was not created");
       }
+
 
       // Update cache
       await cacheFolder(data.id, data);
@@ -73,6 +107,25 @@ export default function FoldersProvider({children} : PropsWithChildren) {
   }
 
   async function deleteUserFolder(folderId: string): Promise<SuccessType | ErrorType> {
+
+    if(networkStatus === 'offline') {
+      const mutation: FolderMutation = {
+        id: Date.now().toString(),
+        type: 'DELETE_FOLDER',
+        folderId,
+      }
+
+      setUserFolders(prev => prev.filter((folder) => folder.id !== folderId));
+
+      await enqueueMutation(mutation);
+      await removeFolderCache(folderId);
+
+      return {
+        error: false,
+        message: 'Folder delete action queued.'
+      }
+    }
+
     const status = await deleteFolder(folderId);
     if(!status) {
       return {
@@ -89,6 +142,37 @@ export default function FoldersProvider({children} : PropsWithChildren) {
   }
 
   async function editUserFolder(newName: string, folderId: string): Promise<SuccessType | ErrorType> {
+
+    if(networkStatus === 'offline') {
+      const tempId = `temp-${Date.now()}`;
+
+      const localFolder: Folder = {
+        id: folderId,
+        name: newName,
+        created_at: Date.now().toString(),
+        user_id: userId?? '',
+      }
+
+      const mutation: FolderMutation = {
+        id: tempId,
+        type: 'UPDATE_FOLDER',
+        folderId,
+        payload: { name: newName },
+      }
+
+      setUserFolders(prev => (
+        prev.map((folder) => folder.id === folderId ? localFolder : folder)
+      ))
+
+      await enqueueMutation(mutation);
+      await cacheFolder(folderId, localFolder);
+
+      return {
+        error: false,
+        message: 'Folder update action queued.'
+      }
+    }
+
     const data = await updateFolder(newName, folderId)
     if(!data) {
       return {
@@ -107,9 +191,60 @@ export default function FoldersProvider({children} : PropsWithChildren) {
     }
   }
 
+  async function processFolderQueue() {
+  const queue = await readMutationQueue();
+  if(queue.length === 0) return;
+
+  for (const mutation of queue) {
+    try {
+      switch (mutation.type) {
+        case 'ADD_FOLDER': {
+          const createdFolder = await createFolder(mutation.payload.name, userId);
+
+          if(!createdFolder) throw new Error();
+
+          await dequeueMutation(mutation);
+          await removeFolderCache(mutation.tempId);
+          await cacheFolder(createdFolder.id, createdFolder);
+
+          break;
+        }
+
+        case 'UPDATE_FOLDER': {
+          const updatedFolder = await updateFolder(mutation.payload.name, mutation.folderId);
+
+          if(!updatedFolder) throw new Error();
+
+          await dequeueMutation(mutation);
+          await cacheFolder(updatedFolder.id, updatedFolder);
+
+          break;
+        }
+
+        case 'DELETE_FOLDER': {
+          const deleteStatus = await deleteFolder(mutation.folderId);
+
+          if(!deleteStatus) throw new Error();
+
+          await dequeueMutation(mutation);
+
+          break;
+        }
+      }
+    } catch (e) {
+      console.log(e);
+      return
+    }
+  }
+}
+
   console.log("Fetched folders:", userFolders)
   useEffect(() => {
-    refreshFolders()
+    async function refreshAppFolders() {
+      await processFolderQueue();
+      await refreshFolders();
+    }
+    refreshAppFolders();
   }, [userId])
 
   return (
